Move pairlists and background tasks API's to separate file
